+++ Otros/Peleas/Peleadores/Luchadores.cs
@@ -9,7 +9,7 @@
         public int vida_actual { get; set; }
         public int vida_maxima { get; set; }
         public byte pa { get; set; }
-        public int pm { get; set; }
+        public byte pm { get; set; }
 
         public int porcentaje_vida => (int)((double)vida_actual / vida_maxima) / 100;
 
@@ -19,9 +19,9 @@
             equipo = _equipo;
         }
 
-        public Luchadores(int _id, bool _esta_vivo, int _vida_actual, byte _pa, int _pm, short _celda_id, int _vida_maxima, byte _equipo) => get_Actualizar_Luchador(_id, _esta_vivo, _vida_actual, _pa, _pm, _celda_id, _vida_maxima, _equipo);
+        public Luchadores(int _id, bool _esta_vivo, int _vida_actual, byte _pa, byte _pm, short _celda_id, int _vida_maxima, byte _equipo) => get_Actualizar_Luchador(_id, _esta_vivo, _vida_actual, _pa, _pm, _celda_id, _vida_maxima, _equipo);
 
-        public void get_Actualizar_Luchador(int _id, bool _esta_vivo, int _vida_actual, byte _pa, int _pm, short _celda_id, int _vida_maxima, byte _equipo)
+        public void get_Actualizar_Luchador(int _id, bool _esta_vivo, int _vida_actual, byte _pa, byte _pm, short _celda_id, int _vida_maxima, byte _equipo)
         {
             id = _id;
             esta_vivo = _esta_vivo;
